# Assignment_12/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Budget Tracker API",
    "version": "1.0.0",
    "description": "API documentation for User, Category, and Transaction endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/transactions": {
      "get": {
        "summary": "Get all transactions",
        "tags": [
          "Transactions"
        ],
        "responses": {
          "200": {
            "description": "List of transactions"
          }
        }
      },
      "post": {
        "summary": "Create a new transaction",
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transactionDetails": {
                    "type": "object",
                    "required": [
                      "transactionId"
                    ],
                    "properties": {
                      "transactionId": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "summary": "Get a transaction by ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction found"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a transaction",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Transaction deleted"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    },
...
  },
...
}